refactor: 상품 목록 data source 네이밍 수정
