+++ app/src/main/java/woowacourse/shopping/data/remote/shopping/RemoteProductDataSourceImpl.kt
@@ -5,18 +5,18 @@ import woowacourse.shopping.data.remote.ApiClient
 import woowacourse.shopping.data.remote.dto.response.ProductDto
 import woowacourse.shopping.data.remote.dto.response.ProductResponse
 
-class RetrofitProductApiService : ProductApiService {
-    private val retrofitProductService: RetrofitProductService =
-        ApiClient.getApiClient().create(RetrofitProductService::class.java)
+class RemoteProductDataSourceImpl : RemoteProductDataSource {
+    private val productApiService: ProductApiService =
+        ApiClient.getApiClient().create(ProductApiService::class.java)
 
     override fun loadById(productId: Long): Call<ProductDto> {
-        return retrofitProductService.requestProductDetail(productId = productId.toInt())
+        return productApiService.requestProductDetail(productId = productId.toInt())
     }
 
     override fun load(
         startPage: Int,
         pageSize: Int,
     ): Call<ProductResponse> {
-        return retrofitProductService.requestProducts(page = startPage, size = pageSize)
+        return productApiService.requestProducts(page = startPage, size = pageSize)
     }
 }
